+++ util/oc_mmem.h
@@ -35,12 +35,14 @@
 #ifndef OC_MMEM_H
 #define OC_MMEM_H
 
+#include <stddef.h>
+
 #define OC_MMEM_PTR(m) (struct oc_mmem *)(m)->ptr
 
 struct oc_mmem
 {
   struct oc_mmem *next;
-  unsigned int size;
+  size_t size;
   void *ptr;
 };
 
@@ -51,22 +53,22 @@ void oc_mmem_init(void);
 #ifdef OC_MEMORY_TRACE
 
 #define oc_mmem_alloc(m, size, pool_type)                                      \
-  (int)_oc_mmem_alloc(__func__, m, size, pool_type)
+  _oc_mmem_alloc(__func__, m, size, pool_type)
 #define oc_mmem_free(m, pool_type) _oc_mmem_free(__func__, m, pool_type)
 
 #else /* OC_MEMORY_TRACE */
 
 #define oc_mmem_alloc(m, size, pool_type)                                      \
-  (int)_oc_mmem_alloc(m, size, pool_type)
+  _oc_mmem_alloc(m, size, pool_type)
 #define oc_mmem_free(m, pool_type) _oc_mmem_free(m, pool_type)
 
 #endif /* !OC_MEMORY_TRACE */
 
-int _oc_mmem_alloc(
+size_t _oc_mmem_alloc(
 #ifdef OC_MEMORY_TRACE
   const char *func,
 #endif
-  struct oc_mmem *m, unsigned int size, pool pool_type);
+  struct oc_mmem *m, size_t size, pool pool_type);
 
 void _oc_mmem_free(
 #ifdef OC_MEMORY_TRACE
